pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "odoo-skills-dev-tools"
version = "1.0.0"
description = "Development tooling for the odoo-skills Claude Code plugin"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
odoo-skills-verify = "tooling.cli:verify_main"

[tool.setuptools.packages.find]
include = ["tooling*"]
